+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Initial integration with HID / Origo sdk for mobile access [#2921](https://github.com/rokwire/illinois-app/issues/2921).
+
 ## [5.0.6] - 2023-05-29
 ### Added
 - Add survey creation tool [#3295](https://github.com/rokwire/illinois-app/issues/3295).